Update schema structs and use source2gen license
Not the full header though as we won't need all of it

+++ src/cs2_sdk/schema.h
@@ -39,8 +39,9 @@
 #include "stdint.h"
 #undef schema
 
-struct SchemaKey {
-	int16_t offset;
+struct SchemaKey
+{
+	int32 offset;
 	bool networked;
 };
